Route handlers are now RESTful
