+++ server/bootstrap/routes/DELETE.js
@@ -1,0 +1,9 @@
+/**
+ * RESTful:
+ * Connects DELETE requests with corresponding Controllers and Actions.
+ * Use to remove existing entries, e.g. in a database or cache.
+ */
+
+exports = module.exports = function(app){
+
+};
