+++ server/bootstrap/routes/POST.js
@@ -1,0 +1,9 @@
+/**
+ * RESTful:
+ * Connects POST requests with corresponding Controllers and Actions.
+ * Use to create new entries, e.g. in a database or cache.
+ */
+
+ exports = module.exports = function(app){
+
+ };
