+++ server/bootstrap/routes/PUT.js
@@ -1,0 +1,9 @@
+/**
+ * RESTful:
+ * Connects PUT requests with corresponding Controllers and Actions.
+ * Use to update existing entries, e.g. in a database or cache.
+ */
+
+ exports = module.exports = function(app){
+
+ };
